refactor: location type 이동
