+++ src/types/common/index.ts
@@ -2,16 +2,3 @@ export interface IPostResponse {
   success: boolean;
   message: string;
 }
-
-export interface ILocation {
-  latitude: number;
-  longitude: number;
-}
-
-export interface IKakaoSearchResponse {
-  id: number;
-  place_name: string;
-  road_address_name: string;
-  x: string;
-  y: string;
-}
